Switched to from port 631 to port 10631
This allows to install the snap on a classic system which is already
running CUPS and cups-browsed. The snap's CUPS is then running on port
10631, independent of the system's CUPS on standard port 631.

The snap's cups-browsed is attached to the snap's cupsd while the
system'e cups-browsed stays attached to the system's cupsd.

This way two printing stacks (one of the system, the other of the snap
are running in parallel and the cups-browsed of each of the two
printing stacks is making available the printers which the other stack
is sharing.