Fixes in the cupsd.conf creation during startup
Made sure that there are the following lines in the beginning of the
file

```
Listen /var/snap/cups/x*/var/run/cups.sock
Port 10631
```

and no further `Listen` and `Port` lines.

Added `Allow @LOCAL` to the `<Location />` section, so that all
machines in the local network have access to the snap's cupsd. This is
also needed for the local system's cups-browsed to access the snap's
printers, as Avahi advertises the snap's printers with the system's
network host name and not with "localhost". Without this change the
snap's shared printers are only accessible with "localhost:10631"
URIs.