Removed use of MuPDF and removed use of the daemon-dbus plug
For now we are switching over to Poppler as the PDF interpreter as we
want to use Poopler anyway, because Poppler is needed to support
printing of test pages and flattening filled PDF forms.

Classic systems do not provide a daemon-dbus slot, so we have to
remove the daemon-dbus plug to use the snap on classic systems.