Remove var/spool/cups/tmp from stage directory
The var/spool/cups/tmp which CUPS creates with "make install¨ is not
actually needed in the snap as one cannot write in the
squashfs. Instead, we had already created an alternative directory by
the run-cupsd script.

As CUPS creates the directory with unusual permissions it offends the
upload of the snap to the Snap Store. Therefore we simply remove it.