Build Ghostscript with only the actually needed output devices
The Ghostscript in the CUPS snap sandbox is only used for the very limited tasks which CUPS and cups-filters require and not a universal PostScript/PDF interpreter and printer driver tool.

Therefore we build it here with only the actually needed output devices which are

  - pdfwrite          PDF output for PDF printers and Printer Applications
  - ps2write          PostScript output for PostScript printers
  - cups              CUPS Raster output for filters
  - pwgraster         PDF output for PDF printers and Printer Applications
  - pxlmono, pxlcolor PCL-XL output for printers

These are the formats needed for driverless printing, including some legacy printers, for Printer Applications, and for filters applied after Ghostscript.