Added DEBUG build option and worked arround cupsGetPPD3() bug
This commit is result of finding a bug in the cupsGetPPD3() of libcups. The function does not work when using CUPS with another port than the standard 631. I falls back to 631 in such a case, accessing the wrong CUPS daemon if there is another one.

This resulted in cups-browsed releasing the queues it created when a job was printed on them, as the check of the integrity of the queue always failed due to the fact that the PPD could not get loaded. The same happened when removing the queues during shutdown, resulting in the queues no removed, but their cups-browsed bit removed, making them breaking subsequent cups-browsed sessions.

This commit contains the following changes:

- cups-browsed is started with the CUPS_SERVERROOT environment variable set, so that cupsGetPPD3() grabs the PPD directly from the file system, as a workaroung for the bug
- General temporary file directory created and TMPDIR environment variable set, as it is needed by the cupsGetPPD3() function
- Support for CUPS DEBUG build: Appropriate, commented out lines are added to snapcraft.yaml, run-cupsd, and run-cups-browsed. Uncommenting them builds CUPS in DEBUG mode and lets cupsd and libcups log very verbosely in $SNAP_DATA/var/log/debug_log and $SNAP_DATA/var/log/cups-browsed-debug_log.