Updated cups-filters to version 1.27.0
In this version of cups-filters cups-browsed does not need to know the port number of the CUPS daemon it is attached to when it connects via domain socket. This makes the snap reliably working both with CUPS on the standard port 631 and with CUPS on an alternative port.

There is no need any more for the run-cups-browsed script to determine on which port the snap's CUPS is listening. The code used for that got removed now.