Do not re-create the config files on every restart
Before, the config files cups-files.conf, cupsd.conf, fonts.conf, and cups-browsed.conf were re-created on each start of the daemons, aminly to get the non-standard paths needed for the snap into the files. Problem was that a user could not do any modifications on the configuration.

Now we create config files only if they are missing and otherwise we only edit in the paths on each start of the daemons, so that most path-independent settings can get modified by the user without getting lost on restart or update of the snap.