Replaced the Python script to check the CUPS port by a C program
To check whether there is already a system-supplied CUPS daemon running on port 631 a simply Python script was used as a preliminary solution. As Python scripts are rather resource-intensive on startup, needing to load many libraries it is not recommended to use them in a service which is started during boot. Therefore the script gets replaced by an equivalent C program now.