Added patch for cupsd to check whether a client plugs "cups-control" interface
In our CUPS Snap, we want to control by the interface the a client
Snap is using to connect to CUPS whether we allow administrative tasks
(create a new printer, cancel all jobs, ...) or only user tasks
(listing printers and capabilities, printing, cancel the requester's
own jobs, ...).

If the client is a Snap and it does not plug the "cups-control"
interface, we deny administrative tasks, even if the user who called
the client is root or member of the CUPS admin group. Only if it plugs
to "cups-control" we allow it to trigger administrative tasks. If the
client is not s Snap or not connecting through the domain socket, we
do not apply any extra restriction for administrative tasks.

This way we can replace the former "cups-control" interface in snapd
by two new interfaces, "cups" which allows access to the CUPS domain
socket and to the CUPS notification D-Bus services, and "cups-control"
which principally has the same access rights as "cups" but with our
patch CUPS allows administrative tasks when using this one.

The "cups" interface, when plugged by an application Snap, will get
automatically connected by default, to make it easy for a Snap
developer to let an application print. The "cups-control" interface
will not connect automatically by default when used in a Snap, as it
is already the case with the current "cups-control". This makes it
easy to get an app which prints into the Store but difficult to get
one in which performs an attack through the printing sub-system,
especially with the background that the first user in an Ubuntu
installation is always in the CUPS admin group.

To implement this we identify at first whether an inquiry from a
client coming through a domain socket is administrative or not. We
consider the task administrative when it is only allowed due to the
fact that the user is in one of the groups listed in the "SystemGroup"
directive in cups-files.conf. This is the case when the task is
authorized via a "require user @System" line in the task's "limit"
section of the policy in cupsd.conf.

In this case we hook in another check, polling the PID of the client
process via peer creds and checking whether this PID is under AppArmor
confinement and whether the confinement comes from a Snap, and from
which Snap.

After determining the name of the client Snap via AppArmor we use the
snapd-glib library functions to first poll the general data record of
the Snap and with this check whether it is under classic
confinement. If so, we allow the operation.

If our client Snap is NOT under classic confinement, we poll its
connections but only its plugging of "cups-control". If we find it
plugging "cups-control" we allow the operation, otherwise we deny it.

So our patch applies the same criteria to allowing administrative CUPS
tasks as PulseAudio applies to allow audio recording. See the patch
for PulseAudio as Snap here:

https://git.launchpad.net/ubuntu/+source/pulseaudio/tree/debian/patches/0700-modules-add-snappy-policy-module.patch

See also the Snapcraft forum thread discussing this change:

https://forum.snapcraft.io/t/interface-request-cups-control-on-cups-snap-and-including-d-bus/