Removed patch libcups-do-not-read-cupsd-conf-from-local-file.patch
This patch is not needed any more, as now we are patching the CUPS
build system so that libcups loads configuration files from the Snap's
SEVERROOT and not from the standard one (we needed this to make our
libcups using client.conf).

Now our libcups will not try to load a cupsd.conf of a conventionally
installed CUPS any more and choke on this due to the Snap's
sandboxing.