Drop privileges of filters and backends using the "snap_daemon" user
CUPS drops the privileges of filters, backends, and other auxiliary
processes by running them as the system user "lp". In a Snap this
kind of system users cannot be used. The only unprivileged system
user available in a Snap is "snap_daemon" and it is also to be used
to run processes unprivileged. So we use "snap_daemon" as the user
under which to run filters and backends ("User snap_daemon" and
"Group snap_daemon" in cups-files.conf).

Also delete the "configure" script after applying all patches and
before running "autoconf", to assure that "autoconf" actually rebuilds
the "configure" script.