Let CUPS use "`setgroups(0, NULL)`" to drop supplementary group IDs
CUPS uses "`setgroups(1, &gid)`" to drop supplementary group ID lists
from auxiliary processes, like filters (with `gid` being the group
under which the process is supposed to run). This does not work in a
Snap. Therefore we replace it by "`setgroups(0, NULL)`". CUPS does not
use this in the first place as it is not portable.