CUPS admin group: Check "lpadmin" and "adm", otherwise go root-only
In standard CUPS users in the group "lpadmin" (and also root) can do
administrative tasks, like creating print queues or deleting other
user's jobs.

As a Snap cannot create application-specific system users or groups,
we check whether "lpadmin" already exists and if not we use the
general admin group "adm" instead. If this also does not exist, we
simply only allow root to do administrative tasks.

This is defined by the "SystemGroup" directive in cups-files.conf.