Build CUPS without systemd support, this way cupsctl works correctly
When building CUPS with all build options at their default values
changing the CUPS configuration with the "cupsctl" command does not
work. "cupsctl" shows the configuration summary correctly when being
started without argument, but when supplying any configuration option,
cupsd.conf gets appropriately modified, but cupsd does not get started
after the change.

Such a behavior was already there, back in CUPS 2.2.12
(https://github.com/apple/cups/issues/5640) and there was done a
correction on the on-demand use handling of cupsd on systems using
systemd as PID 1.

So I came to the idea to test with systemd support turned off
altogether ("./configure --disable-systemd") as inside the Snap we run
CUPS permanently and not on-demand anyway. This made cupsctl working
correctly again, restarting cupsd after applying the configuration
change. So we are going with "--disable-systemd" now.