When checking for an already running CUPS, check also if lpstat errors
To check for an already running classically installed CUPS before starting
the Snap's CUPS, we do a "lpstat -r" call on the standard domain socket
and check this way whether a CUPS daemon is running on it.

We did not check though whether the "lpstat -r" call errors (exits with a
non-zero status). Now we also do this check and if lpstat errors, we also
let the Snap's CUPS run on the alternative domain socket, as if there was
a CUPS running on the standard socket.

In addition, we do not suppress the first "lpstat -r" call's (the call
to check whether lpstat errors) console output, to let it go into the
Snap's logs for debugging.