Put groups listed as "SystemGroups" in cups-files.conf in correct order
cups-files.conf lists the groups users have to be member of to be
allowed to do administrative tasks on CUPS (create queues, delete
someone else's jobs, ...) in its "SystemGroups" directive. Simply for
the decision whether a user gets admin permission or not the order of
the listed group names does not matter, but CUPS also gives to some
directories or files an admin group as group ownership and here it
uses always the first in the list, so here put the actual admin group
first and the fallback root last, as in the classic .deb installation.