Patched CUPS to set permissions/ownerships of certs directory suitable for Snap
Inside a Snap executables are run without CAP_DAC_OVERRIDE capability,
meaning that processes running as root also have to obey permissions
and ownerships of files and directories, in contrary to the standard
behavior where processes as root can always access.

CUPS is written assuming that the standard behavior applies, so small
changes, like this one are needed.

See also

https://forum.snapcraft.io/t/snapped-daemon-running-as-root-cannot-create-file-in-directory-with-odd-ownerships-permissions/