Corrected the CUPS patch to set ownerships/permissions for the certs directory
The snapped CUPS daemon cannot assign arbitrary ownerships to files,
probably due to another missing capability. In our cases it cannot set
the group ownership of the certs directory to the CUPS admin group
("lpadmin" or "adm") as the classically installed CUPS does.

By trying out I have found out that "root" as group ownership also
works.  Itseems that only CUPS itself needs full access to the certs
directory and CUPS itself runs as root. For the group ownership it
seems only be important to not be "snap_daemon" ("lp" in classically
installed CUPS) so that unprivileged sub-processes cannot access.