Removed unneeded documentation and development files from the Snap
Development files, like C header files (include/ directories), and
most documentation files, especially of Debian packages installed into
the Snap to fulfill the dependencies of the provided software and also
the man pages, are removed in the priming step of building the
Snap. The files cannot be removed already in the staging step, as
there the C headers have to be available to build subsequent parts
inside the Snap.

The CUPS documentation for the online help system in the CUPS web admin
interface is moved from $SNAP/share/doc/cups to $SNAP/share/cups/doc to
easily remove the whole $SNAP/share/doc/ on priming to get rid of all
the unneeded documentation files.