Updated cups-filters to version 1.27.5
In this version cups-browsed sets up local queues pointing to "remote"
CUPS printers from a second instance of CUPS on the local machine
always unencrypted (ipp://...), as there are problems with encrypted
communication (ipps://...) between the CUPS servers.

if one runs this Snap on a machine where there is already a
classically installed cupsd running and so this Snap runs its own CUPS
on an alternative domain socket and port, one can print on queues
which the Snap's cups-browsed has created and which point to queues of
the classically installed CUPS instance.