Renamed the CUPS Snap to "cups"
The CUPS Snap on OpenPrinting will be the official Snap, as we cannot
expect that CUPS’ upstream, Apple, will make a Snap of CUPS and also
CUPS on any non-Apple system is used with cups-filters, which is an
OpenPrinting project and it is based on the parts of CUPS which Apple
has spun out due to the fact that they are not needed by Mac OS X and
so Apple did not want to maintain them. Therefore it should be no
problem to name our CUPS Snap simply “cups”.

The GitHub repository is now named “cups-snap” so that the visitor of
our GitHub will not think that this is a fork of CUPS but rather the
set of files needed to build the CUPS Snap.