Make Ghostscript build on ppc64el and arm64
The libpng shipping with Ghostscript does not build on ppc64el and
arm64 without disabling certain optimizations. For ppc64el this is
already solved in the Ghostscript GIT and part of the upcoming 9.53:

http://git.ghostscript.com/?p=ghostpdl.git;a=commitdiff;h=1c9bd8e0aa4
https://bugs.ghostscript.com/show_bug.cgi?id=702236

For arm64 there is a similar problem but Ghostscript developers seem
not yet aware of it. A hint for a solution can be found in another
project though:

https://github.com/imagemin/optipng-bin/issues/97