Added "cups-config" plug for the "cupsd" app to do migration from classic CUPS
The "cups-config" plug allows to the "cupsd" app to read the /etc/cups
directory with all its files to find out about the configuration and
print queues of a classic CUPS installation which is about to be
replaced by the CUPS Snap.

The code for doing the migration is not yet implemented but the plug
we already ad here as it needs automatic connection and currently we
have posted a request for all needed automatic interface connections
here

https://forum.snapcraft.io/t/request-cups-snap-cups-auto-connection-to-avahi-control-raw-usb-cups-control-and-system-files-interfaces/

This way we can include this plug.