libcups did not resolve DNS-SD-based URIs correctly when from loopback interface
If a DNS-SD-advertised printer is a local-only (only via loopback,
"lo" network interface) service (Printer Application or IPP-over-USB)
libcups does resolve its DNS-SD-service-name-based URI correctly,
using the local system's network host name, through which the
local-only service is not accessible.

This patch makes the network interface checked and if it is "lo",
"localhost" is used as the host name.