update snapcraft.yaml for existing and new cups interfaces
https://github.com/snapcore/snapd/pull/8920 is adjusting the
cups-control interface so app snaps may provide it. The permanent slot
policy adds the various accesses that the 'system-config' and
'cups-domain-socket' system-files interfaces provided, so have cupsd
slot the cups-control interface.

In addition, the PR adds a new 'cups' interface that cupsd may also
slot. The intention of this interface is that the cups snap will slots
both cups-control and cups where both interfaces provide access to the
cupsd socket for plugging snaps such that when mediation patches are
applied, cupsd can query snapd to see if the connecting process is a
snap, and if it is determine if it is is confined, and if it is, check
if it has the cups-control interface connect before granting access to
admin functionality.

With the above in place, adjust all non-cupsd commands/services within
the snap to plugs cups-control instead of cups-domain-socket. Also
remove all references to system-config.

Finally, adjust cupsd to plugs mount-observe, cups-browsed to plugs
network-manager-observe and gs to plugs etc-cups.

With the above in place, cupsd and its utilities should have all the
required access to perform their duties. The snap still has cupsd
plugging 'snapd-control'; this will need to change to use the upcoming
snap query API.

Snaps may use:
* 'plugs: [ cups-control ]' to have access to printing and admin
  functionality
* 'plugs: [ cups ]' to have access to printing for slot providers with
  the snap mediation patches

For systems with the cups snap installed (eg, Ubuntu Core or systems
with the deb/rpm/etc removed), manual connection is now:

$ sudo snap connect foo:cups cups:printing
$ sudo snap connect foo:cups-control cups:admin