Added support for new "snapctl is-connected" functionality
A security feature which we were long after is to check on
administrative requests to the CUPS daemon (create/modify print
queues, delete somebody else's jobs, ...) whether if they come from a
fully-confined Snap this client Snap plugs "cups-control" to our
Snap's "admin" slot. Only than these requests are accepted.

With the Snap Store allowing automatic connection of the "cups"
printing interface to the "printing" slot of this Snap by default and
not allowing automatic connection of "cups-control" it gets more
difficult to post Snaps in the Snap Store with malicious
administrative action on CUPS.

On a first approach we had to plug to the dangerous interfaces
"mount-observe" and "snapd-control" to do this using libapparmor amd
libsnapd-glib functions, giving full control on snapd, an approach for
which the Snap Store team never will allow exceptions for
automatically connecting the appropriate interfaces.

So we ended up to extend the functionality of snapd and its snapctl
utility. The call

snapctl is-connected --apparmor-label=snap.cups.lpadmin admin

checks whether the Snap specified by the given AppArmor profile name
connects to our "admin" slot, using its "cups-control" plug. In this
case the function exits with 0 (so we grant permission), if not with 1
(we deny permission).

It also recognizes whether the client is a classic Snap, returning 10
or no Snap at all returning 11. In both these cases we always grant
permission. This assures that the CUPS Snap also works with clients
installed via DEB or RPM for example.

Observation: For the admin utilities coming with this Snap to work,
the Snap has to bite into its own tail, via the command:

sudo snap connect cups:cups-control cups:admin

We are dropping the "mount-observe" and "snapd-control" from cupsd
as they are not needed any more.