Updated plugs of the command line tools to distinguish between administrative or not
Let only the command line tools for administrative tasks plug
"cups-control" and the ones which are only for printing and checking
status only plug "cups".

Also remove unneeded "avahi-control" plugs from many command line
tools.

Updated README.md because also the "cups" plug of the CUPS Snap needs
to be connected to the "printing" slot of the very same Snap for the
non-administrative command line tools to work.