Joined all CUPS upstream Snap support in single patch, do not use "snapctl"
All patches on CUPS to make it running in a Snap are now applied
upstream and until there is a CUPS release with these changes they are
all joined into a single patch now.

In addition, the check whether a client requesting an administrative
task is a Snap and whether it plugs "cups-control" is not done by the
call of the external executable "snap-control" any more but by the new
library function snapd_client_run_snapctl2_sync() of
libsnapd-glib. This scales better if there are very many inquiries to
cupsd in parallel.

As there is no release of libsnapd-glib with this function yet, we load
the library from the upstream GIT for our Snap build instead of using
the package of the core distribution.