Eliminated the last sed use from the CUPS build
The "sysconfdir" (where configuration files go) for the CUPS Snap was
still set by modifying the a .m4 file which is used to build the
./configure script. This is now replaced by the ./configure option
"--sysconfdir=...".

As this change leads to the sample/default configurations files, from
which the initial configuration files are derived, to be installed to
an awkward place, we add an "organize:" entry to correct this.