THe "usb" CUPS backend needs to run as root, set permissions appropriately
CUPS upstream sets world-executable permissions to the "usb" backend,
so that it runs as unprivileged system user and not as root. In most
distributions this works by setting the raw USB devices for printers
to "lp" group ownership via UDEV rules.

In the Snap this is not possible, the Snap has no influence on which
groups are available and to which group the raw USB devices are set,
or whether the host distribution is even nit using this group concept
and running the "usb" backend as root instead. So we need the more
universal solution of running the "usb" backend as root for our Snap,
this always works.

The implementation in this commit is preliminary, a solution in the
build system of CUPS upstream is looked into and preferred.