Re-defined plugs and slots to have standard names
The slots of the CUPS Snap where client Snap's "cups" and
"cups-control" are supposed to plug to have the non-standard names
"printing" and "admin" whereas standard is to use the same name as the
plugs, "cups" and "cups-control" in our case. This we are doing her
now.

As the included utilities also plug "cups" and "cups-control", as
external client Snaps we get a name clash. Therefore we define a new
plug named "cups-internal" which is not more than an alias name of
"cups-control" and let all the included tools plug "cups-internal"
instead of "cups" and "cups-control". Here is also important to say
that "cups-control" allows both printing and administration so one
interface instead of two is good enough here.

We also build CUPS with the correct slot name for the "snapctl" calls,
using the "--with-cups-control-slot=cups-control" option on
"./configure".

For the included tools we only need this connection now:

sudo snap connect cups:cups-internal cups:cups-control

See

https://forum.snapcraft.io/t/request-cups-snap-cups-auto-connection-to-of-cups-cups-control-to-cups-admin-and-also-of-the-network-manager-observe-interface/