Removed the "slots: [cups-control, cups]" from the "cupsd" app definition
This way the slots get associated with all apps of the Snap, giving
the utilities access to the slots.

The utilities now even work when their plug ("cups-internal") is not
connected to cupsd's "cups-control" slot, but we cannot remove the
"cups-internal" plug from the utilities though as otherwise they
cannot access an external CUPS daemon, outside this Snap.