Added new proxy mode, to pass on client Snap's jobs to system CUPS
Following the discussion here

https://github.com/snapcore/snapd/pull/10023#issuecomment-799425764

https://forum.snapcraft.io/t/handling-of-the-cups-plug-by-snapd-especially-auto-connection/23419/3?u=till.kamppeter

I am implementing Ian Johnson's suggestion (see second link) to solve
the problem of a confined application Snap only being able to print
but not to be able to do administrative tasks on the system's CUPS
when the application Snap only plugs "cups" (which auto-connects by
default) and not "cups-control" (which only auto-connects with
permission of the Snap Store team).

If the system's CUPS is this CUPS Snap or a the CUPS package of a
current Ubuntu distribution, there is a Snap mediation in cupsd which
checks if an incoming request is administrative, whether it comes from
a confined Snap or not and in the former whether it plugs
"cups-control", before accepting it. So in these cases the security
concept works amd allows arbitrary developers to upload application
which print to the Snap Store and users are protected against
administrative tasks performed on their CUPS by these applications,
even if the user is in the "lpadmin" group (usually the first created
user on Ubuntu is).

Now if we take an old Ubuntu or a non-Ubuntu distribution (everything
where one can install snapd on, usually everything where PID 1 is
systemd) then we have a CUPS without Snap mediation, so we restrict
the "cups" interface to the domain socket of this CUPS Snap and make
the installation of a Snap which plugs "cups" install this CUPS Snap
as a dependency (default-provider).

But forcing the use of the CUPS Snap for printing from snapped
applications and continuing with the original system's CUPS for
unsnapped applications causes user confusion and users cannot print
with their classic (often proprietary) printer drivers.

So to make everything transparent and consistent, we let this CUPS
Snap run as a proxy for the system's classic CUPS when it is installed
on a system which has a classic CUPS installed (determined by the
presence of an /etc/cups/cupsd.conf file). In this case it runs a
special auxiliary daemon (cups-proxyd, part of this Snap, NOT
cups-browsed) which mirrors all ths queues of the system's classic
CUPS to the CUPS of this Snap and on the Snap configures them tompass
the jobs through to the system's classic CUPS unfiltered, so that
drivers are used there.

So an application Snap plugging "cups" cannot touch the system's CUPS
at all and any administrative tasks issued by it are blocked by the
Snap mediation of the snapped CUPS (so it also cannot mess up the
snapped CUPS). The snapped CUPS only grabs queues and copies PPDs (to
make the options available to the client) from the system's CUPS and
sends jobs to the system's CUPS. Only an application Snap plugging
"cups-control" interacts directly with the system's CUPS.

If there is no classic CUPS (no /etc/cups/cupsd.conf), then the Snap
acts as a standalone system CUPS and also starts cups-browsed. It
listens on the standard domain socket and port then and so all
(snapped and unsnapped) applications print on this CUPS. As the
snapped CUPS has Snap mediation it does not need to be protected by
another proxy CUPS.

If you want to use the old way of this Snap's CUPS running as an
independent second CUPS daemon in parallel to the system's classic
CUPS, as this Snap did before when there is a classic CUPS already,
you need to simply run

sudo touch /var/snap/cups/common/no-proxy

to block the proxy mode. Remove the file again to enable the proxy
mode again. You need to stop the CUPS Snap and start it again for this
to take effect:

sudo snap stop cups
sudo snap start cups

Note that this mode is NOT RECOMMENDED FOR PRODUCTION. It is not
tested much and two independent instances of cupsd are confusing for
users.