In stand-alone mode let CUPS listen on BOTH standard and Snap domains
This way CUPS is listening on the Snap's domain
/var/snap/cups/common/run/cups.sock in all three modes (stand-alone,
proxy, and parallel), so that snapped client applications which want
to print can ALWAYS use the Snap's domain and so always print through
Snap-mediated CUPS, which is the standard CUPS when there is no
classically installed CUPS or the proxy to the classic CUPS if there
is a classically installed CUPS.

Classically installed applications always print through the standard
domain /run/cups/cups.sock, using the classically installed CUPS if
present or the snapped CUPS in stand-alone mode if this is the only
CUPS on the system.

cupsd can listen through any amount of different channels, even more
than one domain socket. One simply puts any amount of "Listen" and
"Port" lines in cupsd.conf to define the desired channels.