core20 has mdns4_minimal now, so remove the workaround
In snapcraft.yaml removed the mdns-lookup: part for installing
libnss-mdns and creating the corrected nsswitch.conf and removed the
layout: directive to apply the new nsswitch.conf.

core20 now supports mdns4_minimal out-of-the-box in the officially
released version and the layout: workaround seems to break on systems
where nsswitch.conf is a symlink.

See https://github.com/OpenPrinting/ghostscript-printer-app/issues/3