Switched cups-filters over to GIT master (2.x generation)
The second generation of cups-filters is under heavy development and
shortly before its first release. Its main purpose is support for the
New Architecture, PPD-less all-IPP printing.

Now we are switching over in the CUPS Snap, not only to test more
parts of it in a Snap environment (the filter functions already have
Snap experience in the Printer Applications retro-fitting classic CU{S
drivers) but also to fix some known bugs in cups-filters 1.x:

- Raster-only printers do not print as destination printer of a queue
  created by cups-browsed
  (https://github.com/OpenPrinting/cups-filters/issues/426)
- Page could be de-centered if non-printable margins of printer are not
  symmetric
  (https://github.com/OpenPrinting/cups-filters/commit/31dfcae)
- Adjust cups-browsed's namimg of locally created queues with CUPS'
  naming of temporary queues to avoid duplicates
  (https://github.com/OpenPrinting/cups-filters/commit/3d9c480)
- Many other fixes

Most, but not all fixes will be backported to cups-filters 1.x if
applicable, but we sooner or later want to switch over to cups-filters
2.x in the CUPS Snap anyway, and the sooner we do so, the more chances
we have to find further bugs.