Added versioning
Added a scriptlet to the "cups" part in snapcraft.yaml to do an
automatic versioning. The use of the scriptlet is triggered by the
"adopt-info: cups" entry in snapcraft.yaml.

It follows a versioning policy like this (which is very similar to the
one of Debian packages):

- Base version is the upstream version of the included CUPS. Revisions
  which keep the included CUPS version but do other changes, like
  build method patches, the other included software packages, like
  Ghostscript, cups-filters, QPDF, ... increase the added package
  release number, which is added to the upstream version number with a
  dash. Current version number is 2.4.1-1.

- CUPS is always pulled from the GIT. To grab a release, "source-tag:
  v2.4.1" is added, with the release tag of the GIT repository of
  CUPS. In this case the upstream version number of the Snap is the
  onr of the included cUPS.

- To grab a GIT snapshot between releases, the "source-tag: ..." line
  is commented out. to get the top of the GIT repository. In this case
  the upstream version number is the version number of the last CUPS
  release with added "+gitN.gXXXXXXXXX", where N is the number of
  commits with happened after the release and XXXXXXXXX is the commit
  ID. N is important to make the version number of a newer commit
  actually considered newer.

- If a new Snap is built with the same upstream version of the newest
  Snap currently in the Snap Store, the package release number is
  automatically increased by 1.

- Snaps built of released CUPS versions are considered stable and get
  the "stable" grade and Snaps based on GIT snapshots are considered
  development snapshots and get the "devel" grade.

- Grade and package release number can get manually overridden.

This kind of versioning is also planned to be used by the other Snaps
at OpenPrinting. It is especially suitable also if the Snap packaging
is a repository separate from teh underlying upstream project.

The versioning is especially introduced now to do a first test of
promoting a CUPS Snap version to stable, to be able to test the "cups"
snapd interface and the auto-installing of the CUPS Snap as a
dependency of client Snaps plugging "cups" needed for the interface to
work.