Added cups-socket-directory definition to "cups" slot
We add the "cups-socket-directory: $SNAP_COMMON/run" to the "cups"
slot definition, so that snapd (version 2.55 or newer) mounts this
directory to /var/cups of clients which plug the "cups"
interface. This way the CUPS socket which is placed in
$SNAP_COMMON/run gets available to the client.

snapd will also set the environment variable
"CUPS_SERVER=/var/cups/cups.sock" on the client, so that the Snap's
CUPS socket will get actually used.