Added dry-run mode to cups-proxyd
Added a dry-run mode, which only logs tha appearing and disappearing
printers for the system's cupsd got added. To use this mode only one
cupsd has to be supplied on the command line ad this one is considered
the system's cupsd.

This mode is for development and debugging, not only of cups-proxyd
itself but also of CUPS for example.