snapcraft.yaml: Prevent cups-proxyd binary from source tree leak into Snap
If for development, debugging, testing, ... the cups-proxyd executable
is built locally by running "make" in the cups-proxyd/ directory in
the source tree of the CUPS Snap and afterwards one builds the Snap
via "snapcraft snap" in the main directory of the source tree, the
locally built executable gets copied in the Snap and no new executable
compiled in the Snap's core20 build environment, leading to a Snap
containing an executable which does not work inside the Snap due to a
GLIBC mismatch, giving an error like:

    $ snap run --shell cups.cups
    $ cups-proxyd
    cups-proxyd: /lib/x86_64-linux-gnu/libc.so.6: version `GLIBC_2.34'
        not found (required by cups-proxyd)

To prevent this, we now run "make clean" before "make" when building
the cups-proxyd part. This way any leaked in binaries will be removed
and new, correct once get created.